nobroker_playwright/requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
jupyter
joblib
requests
pytest-playwright
pytest-reporter-html1
selectolax
pymongo
python-dotenv